refactor: TripService 책임 과다로 인한 클래스 분리
