+++ backend/src/main/java/kr/ac/hs/RandomTrip/trip/controller/TripController.java
@@ -3,6 +3,9 @@ package kr.ac.hs.RandomTrip.trip.controller;
 import java.util.List;
 
 import kr.ac.hs.RandomTrip.trip.service.DirectionService;
+import kr.ac.hs.RandomTrip.trip.service.PlaceSearchService;
+import kr.ac.hs.RandomTrip.trip.service.TripRecommendationService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,45 +19,41 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import kr.ac.hs.RandomTrip.trip.dto.TripRecommendRequestDto;
 import kr.ac.hs.RandomTrip.trip.dto.TripResponseDto;
-import kr.ac.hs.RandomTrip.trip.service.TripService;
 import reactor.core.publisher.Mono;
 
 @RestController
 @RequestMapping("/api/trip")
 @Tag(name = "Trip", description = "Trip API")
+@RequiredArgsConstructor
 public class TripController {
 
-    private final TripService tripService;
+    private final TripRecommendationService tripRecommendationService;
+    private final PlaceSearchService placeSearchService;
     private final DirectionService directionService;
-
-    public TripController(TripService tripService, DirectionService directionService) {
-        this.tripService = tripService;
-        this.directionService = directionService;
-    }
 
     @GetMapping("/random")
     @Operation(summary = "랜덤 관광지 조회", description = "완전한 랜덤 관광지를 추천합니다")
     public ResponseEntity<TripResponseDto> getRandomDestination() {
-        TripResponseDto response = tripService.getRandomDestination();
+        TripResponseDto response = tripRecommendationService.getRandomDestination();
         return ResponseEntity.ok(response);
     }
 
     @PostMapping("/recommend-walk")
     @Operation(summary = "도보 여행 코스 추천", description = "도보로 이동 가능한 여행 코스를 추천합니다")
     public ResponseEntity<List<List<TripResponseDto>>> recommendTripWalk(@RequestBody TripRecommendRequestDto request) {
-        return ResponseEntity.ok(tripService.recommendTrip(request, "도보"));
+        return ResponseEntity.ok(tripRecommendationService.recommendTrip(request, "도보"));
     }
 
     @PostMapping("/recommend-car")
     @Operation(summary = "차량 여행 코스 추천", description = "차량으로 이동하는 여행 코스를 추천합니다")
     public ResponseEntity<List<List<TripResponseDto>>> recommendTripCar(@RequestBody TripRecommendRequestDto request) {
-        return ResponseEntity.ok(tripService.recommendTrip(request, "차량"));
+        return ResponseEntity.ok(tripRecommendationService.recommendTrip(request, "차량"));
     }
 
     @PostMapping("/recommend-transit")
     @Operation(summary = "대중교통 여행 코스 추천", description = "대중교통으로 이동하는 여행 코스를 추천합니다")
     public ResponseEntity<List<List<TripResponseDto>>> recommendTripTransit(@RequestBody TripRecommendRequestDto request) {
-        return ResponseEntity.ok(tripService.recommendTrip(request, "대중교통"));
+        return ResponseEntity.ok(tripRecommendationService.recommendTrip(request, "대중교통"));
     }
 
     // 축제 정보 조회 API - GET 방식으로 변경 (areaCode 사용)
@@ -63,7 +62,7 @@ public class TripController {
     public ResponseEntity<List<TripResponseDto>> getFestivals(
             @Parameter(description = "지역코드 (예: 1-서울, 6-부산, 31-경기도)", example = "1")
             @RequestParam String areaCode) {
-        List<TripResponseDto> festivals = tripService.getFestivalsByAreaCode(areaCode);
+        List<TripResponseDto> festivals = placeSearchService.getFestivalsByAreaCode(areaCode);
         return ResponseEntity.ok(festivals);
     }
 
@@ -87,7 +86,7 @@ public class TripController {
     public ResponseEntity<List<TripResponseDto>> searchDestinationByName(
             @Parameter(description = "검색할 장소 이름", example = "경복궁")
             @RequestParam String title) {
-        List<TripResponseDto> results = tripService.searchPlace(title);
+        List<TripResponseDto> results = placeSearchService.searchPlace(title);
         if (results.isEmpty()) {
             return ResponseEntity.notFound().build();
         }
